moved XP Stuff to their own file for better handling/keeping
